+++ app/src/main/res/navigation/profile_graph.xml
@@ -120,6 +120,9 @@
         <action
             android:id="@+id/action_profileRedesignFrag_to_eventDescRedesignFrag"
             app:destination="@id/eventDescRedesignFrag" />
+        <action
+            android:id="@+id/action_profileRedesignFrag_to_createTeamFragment"
+            app:destination="@id/createTeamFragment" />
     </fragment>
     <dialog
         android:id="@+id/settingsRedesignFrag"
@@ -300,5 +303,10 @@
         android:name="ru.radixit.letsplay.presentation.ui.fragments.tabs.profile.HistoryActiveFrag"
         android:label="frag_history_active"
         tools:layout="@layout/frag_history_active" />
+    <dialog
+        android:id="@+id/createTeamFragment"
+        android:name="ru.radixit.letsplay.presentation.ui.fragments.tabs.profile.teams.CreateTeamFragment"
+        android:label="CreateTeamFragment"
+        tools:layout="@layout/fragment_creat_team"/>
 
 </navigation>
